eco-pack-ai/requirements.txt
pandas
numpy
scikit-learn
xgboost
joblib
psycopg2-binary
sqlalchemy
flask
flask-cors
matplotlib
plotly
streamlit